+++ R/fit_c3_aci.R
@@ -35,6 +35,7 @@ fit_c3_aci <- function(
     hard_constraints = 0,
     calculate_confidence_intervals = TRUE,
     remove_unreliable_param = 2,
+    debug_mode = FALSE,
     ...
 )
 {
@@ -44,6 +45,15 @@ fit_c3_aci <- function(
 
     if (sd_A != 'RMSE') {
         stop('At this time, the only supported option for sd_A is `RMSE`')
+    }
+
+    # Get the replicate identifier columns
+    replicate_identifiers <- identifier_columns(replicate_exdf)
+
+    if (debug_mode) {
+        debug_msg('fit_c3_aci curve identifiers:')
+        cat('\n')
+        utils::str(replicate_identifiers$main_data)
     }
 
     # Define the total error function; units will also be checked by this
@@ -69,6 +79,7 @@ fit_c3_aci <- function(
         cj_crossover_min,
         cj_crossover_max,
         hard_constraints,
+        debug_mode,
         ...
     )
 
@@ -77,7 +88,7 @@ fit_c3_aci <- function(
     required_variables <- list()
     required_variables[[ca_column_name]] <- unit_dictionary('Ca')
 
-    check_required_variables(replicate_exdf, required_variables)
+    check_required_variables(replicate_exdf, required_variables, check_NA = FALSE)
 
     # Assemble lower, upper, and fit_options
     luf <- assemble_luf(
@@ -111,7 +122,7 @@ fit_c3_aci <- function(
         gmc_guess,
         kc_guess,
         ko_guess,
-        100, # cc_threshold_rd
+        100, # cc_threshold_rl
         Wj_coef_C,
         Wj_coef_Gamma_star,
         a_column_name,
@@ -125,10 +136,18 @@ fit_c3_aci <- function(
         rl_norm_column_name,
         total_pressure_column_name,
         tp_norm_column_name,
-        vcmax_norm_column_name
+        vcmax_norm_column_name,
+        debug_mode
     )
 
     initial_guess <- initial_guess_fun(replicate_exdf)
+
+    if (debug_mode) {
+        debug_msg(
+            'fit_c3_aci initial_guess:',
+            paste(initial_guess, collapse = ', ')
+        )
+    }
 
     # Find the best values for the parameters that should be varied
     optim_result <- optim_fun(
@@ -143,6 +162,13 @@ fit_c3_aci <- function(
     # Get the values of all parameters following the optimization
     best_X <- fit_options_vec
     best_X[param_to_fit] <- optim_result[['par']]
+
+    if (debug_mode) {
+        debug_msg(
+            'fit_c3_aci best_X:',
+            paste(best_X, collapse = ', ')
+        )
+    }
 
     # Get the corresponding values of Cc at the best guess
     replicate_exdf <- apply_gm(
@@ -337,6 +363,13 @@ fit_c3_aci <- function(
         }
     }
 
+    if (debug_mode) {
+        debug_msg(
+            'fit_c3_aci outcome:',
+            if (fit_failure) {'failure'} else {'success'}
+        )
+    }
+
     # Include the atmospheric CO2 concentration
     replicate_exdf[, 'Ca_atmospheric'] <- Ca_atmospheric
 
@@ -348,9 +381,6 @@ fit_c3_aci <- function(
 
     # Add a column for the residuals
     replicate_exdf <- calculate_residuals(replicate_exdf, a_column_name)
-
-    # Get the replicate identifier columns
-    replicate_identifiers <- identifier_columns(replicate_exdf)
 
     # Attach identifiers to interpolated rates, making sure to avoid duplicating
     # any columns
@@ -396,18 +426,20 @@ fit_c3_aci <- function(
     replicate_identifiers[, 'Tp_tl_avg']         <- mean(replicate_exdf[, 'Tp_tl'])
     replicate_identifiers[, 'Vcmax_tl_avg']      <- mean(replicate_exdf[, 'Vcmax_tl'])
 
-    # Also add fitting details
-    replicate_identifiers[, 'convergence']         <- optim_result[['convergence']]
-    replicate_identifiers[, 'convergence_msg']     <- optim_result[['convergence_msg']]
-    replicate_identifiers[, 'feval']               <- optim_result[['feval']]
-    replicate_identifiers[, 'optimizer']           <- optim_result[['optimizer']]
-    replicate_identifiers[, 'c3_assimilation_msg'] <- replicate_exdf[1, 'c3_assimilation_msg']
+    # Attach fitting details
+    replicate_identifiers[, 'convergence']           <- optim_result[['convergence']]
+    replicate_identifiers[, 'convergence_msg']       <- optim_result[['convergence_msg']]
+    replicate_identifiers[, 'feval']                 <- optim_result[['feval']]
+    replicate_identifiers[, 'optimizer']             <- optim_result[['optimizer']]
+    replicate_identifiers[, 'c3_assimilation_msg']   <- replicate_exdf[1, 'c3_assimilation_msg']
+    replicate_identifiers[, 'c3_optional_arguments'] <- replicate_exdf[1, 'c3_optional_arguments']
 
-    # Store the results
-    replicate_identifiers[, 'operating_Ci']       <- operating_point_info$operating_Ci
-    replicate_identifiers[, 'operating_Cc']       <- operating_point_info$operating_Cc
-    replicate_identifiers[, 'operating_An']       <- operating_point_info$operating_An
-    replicate_identifiers[, 'operating_An_model'] <- operating_An_model
+    # Attach operating point information
+    replicate_identifiers[, 'operating_Ci']        <- operating_point_info$operating_Ci
+    replicate_identifiers[, 'operating_Cc']        <- operating_point_info$operating_Cc
+    replicate_identifiers[, 'operating_An']        <- operating_point_info$operating_An
+    replicate_identifiers[, 'operating_An_model']  <- operating_An_model
+    replicate_identifiers[, 'operating_point_msg'] <- operating_point_info$operating_point_msg
 
     # Get an updated likelihood value using the RMSE
     replicate_identifiers[, 'optimum_val'] <- if (fit_failure) {
@@ -441,35 +473,37 @@ fit_c3_aci <- function(
     # Document the new columns that were added
     replicate_identifiers <- document_variables(
         replicate_identifiers,
-        c('fit_c3_aci',               'alpha_g',             'dimensionless'),
-        c('fit_c3_aci',               'alpha_old',           'dimensionless'),
-        c('fit_c3_aci',               'alpha_s',             'dimensionless'),
-        c('fit_c3_aci',               'alpha_t',             'dimensionless'),
-        c('fit_c3_aci',               'Gamma_star_at_25',    'micromol mol^(-1)'),
-        c('fit_c3_aci',               'Gamma_star_tl_avg',   'micromol mol^(-1)'),
-        c('fit_c3_aci',               'gmc_at_25',           'mol mol^(-2) s^(-1) bar^(-1)'),
-        c('fit_c3_aci',               'gmc_tl_avg',          'mol mol^(-2) s^(-1) bar^(-1)'),
-        c('fit_c3_aci',               'J_at_25',             'micromol m^(-2) s^(-1)'),
-        c('fit_c3_aci',               'J_tl_avg',            'micromol m^(-2) s^(-1)'),
-        c('fit_c3_aci',               'Kc_at_25',            'micromol mol^(-1)'),
-        c('fit_c3_aci',               'Kc_tl_avg',           'micromol mol^(-1)'),
-        c('fit_c3_aci',               'Ko_at_25',            'mmol mol^(-1)'),
-        c('fit_c3_aci',               'Ko_tl_avg',           'mmol mol^(-1)'),
-        c('fit_c3_aci',               'RL_at_25',            'micromol m^(-2) s^(-1)'),
-        c('fit_c3_aci',               'RL_tl_avg',           'micromol m^(-2) s^(-1)'),
-        c('fit_c3_aci',               'Tp_at_25',            'micromol m^(-2) s^(-1)'),
-        c('fit_c3_aci',               'Tp_tl_avg',           'micromol m^(-2) s^(-1)'),
-        c('fit_c3_aci',               'Vcmax_at_25',         'micromol m^(-2) s^(-1)'),
-        c('fit_c3_aci',               'Vcmax_tl_avg',        'micromol m^(-2) s^(-1)'),
-        c('estimate_operating_point', 'operating_Ci',        replicate_exdf$units[[ci_column_name]]),
-        c('estimate_operating_point', 'operating_Cc',        replicate_exdf$units[[cc_column_name]]),
-        c('estimate_operating_point', 'operating_An',        replicate_exdf$units[[a_column_name]]),
-        c('fit_c3_aci',               'operating_An_model',  replicate_exdf$units[[a_column_name]]),
-        c('fit_c3_aci',               'convergence',         ''),
-        c('fit_c3_aci',               'convergence_msg',     ''),
-        c('fit_c3_aci',               'feval',               ''),
-        c('fit_c3_aci',               'optimum_val',         ''),
-        c('fit_c3_aci',               'c3_assimilation_msg', '')
+        c('fit_c3_aci',               'alpha_g',               'dimensionless'),
+        c('fit_c3_aci',               'alpha_old',             'dimensionless'),
+        c('fit_c3_aci',               'alpha_s',               'dimensionless'),
+        c('fit_c3_aci',               'alpha_t',               'dimensionless'),
+        c('fit_c3_aci',               'Gamma_star_at_25',      'micromol mol^(-1)'),
+        c('fit_c3_aci',               'Gamma_star_tl_avg',     'micromol mol^(-1)'),
+        c('fit_c3_aci',               'gmc_at_25',             'mol mol^(-2) s^(-1) bar^(-1)'),
+        c('fit_c3_aci',               'gmc_tl_avg',            'mol mol^(-2) s^(-1) bar^(-1)'),
+        c('fit_c3_aci',               'J_at_25',               'micromol m^(-2) s^(-1)'),
+        c('fit_c3_aci',               'J_tl_avg',              'micromol m^(-2) s^(-1)'),
+        c('fit_c3_aci',               'Kc_at_25',              'micromol mol^(-1)'),
+        c('fit_c3_aci',               'Kc_tl_avg',             'micromol mol^(-1)'),
+        c('fit_c3_aci',               'Ko_at_25',              'mmol mol^(-1)'),
+        c('fit_c3_aci',               'Ko_tl_avg',             'mmol mol^(-1)'),
+        c('fit_c3_aci',               'RL_at_25',              'micromol m^(-2) s^(-1)'),
+        c('fit_c3_aci',               'RL_tl_avg',             'micromol m^(-2) s^(-1)'),
+        c('fit_c3_aci',               'Tp_at_25',              'micromol m^(-2) s^(-1)'),
+        c('fit_c3_aci',               'Tp_tl_avg',             'micromol m^(-2) s^(-1)'),
+        c('fit_c3_aci',               'Vcmax_at_25',           'micromol m^(-2) s^(-1)'),
+        c('fit_c3_aci',               'Vcmax_tl_avg',          'micromol m^(-2) s^(-1)'),
+        c('estimate_operating_point', 'operating_Ci',          replicate_exdf$units[[ci_column_name]]),
+        c('estimate_operating_point', 'operating_Cc',          replicate_exdf$units[[cc_column_name]]),
+        c('estimate_operating_point', 'operating_An',          replicate_exdf$units[[a_column_name]]),
+        c('estimate_operating_point', 'operating_point_msg',   ''),
+        c('fit_c3_aci',               'operating_An_model',    replicate_exdf$units[[a_column_name]]),
+        c('fit_c3_aci',               'convergence',           ''),
+        c('fit_c3_aci',               'convergence_msg',       ''),
+        c('fit_c3_aci',               'feval',                 ''),
+        c('fit_c3_aci',               'optimum_val',           ''),
+        c('fit_c3_aci',               'c3_assimilation_msg',   ''),
+        c('fit_c3_aci',               'c3_optional_arguments', '')
     )
 
     # Calculate confidence intervals, if necessary